API

Basic Structures

FlowCytometry.FlowCytometryGate — Type

struct FlowCytometryGate

Structure that contains information from a gating process of the data.

Elements:

channels::Tuple{String,String} Channel names of the gate.

polygon::Vector{Tuple{Real,Real}} Points of poligon that defines the gate.

FlowCytometry.FlowCytometryControl — Type

mutable struct FlowCytometryControl

Structure that stores the data coming from control stainings for correcting for spillover. In order for the system to compute properly the spillover matrix, the names of the dictionary must correspond with the names in the channel labels in var.

Elements:

channels::Vector{String} Name of the channels in the experiment.

controls::Dict{String,FlowCytometryExperiment} Dictionary of FlowCytometryExperiments containing the control experiments.

compensationMatrix::Union{Nothing,CompensationMatrix} Spillover matrix inverted.

uns::Dict{String,Any} Dictionary where to store metainformation from spillover algorithm.

FlowCytometry.FlowCytometryExperiment — Type

mutable struct FlowCytometryExperiment

Structure containing a Flow Cytometry experiment and all the processes applyied to it.

Elements:

channels::Vector{String} Name of the channels in the experiment.

X::Matrix{AbstractFloat} Matrix of Cells X Channel of the experiment.

obs::DataFrame Dataframe with all the metainformation of the cells

var::DataFrame Dataframe with all the metainformation of the channels

obsm::Dict{String,Matrix{AbstractFloat}} Dictionary containing transformed matrices of the original data.

layers::Dict{String,Matrix{AbstractFloat}} Dictionary containing Cells X Channel matrices of data that are required to control (e.g. Raw matrix).

gates::Dict{String,FlowCytometryGate} List of Gate and Gate set objects

uns::Dict{String,Any} Dictionary contining all the metainformation of algorithms applied to the data.

Operations with structures

FlowCytometry.removeCells — Function

function removeCells(fcs::FlowCytometryExperiment, s::Vector{Bool})

Remove cells from the dataset and return a copy.

Arguments:

fcs::FlowCytometryExperiment Dataset to be pruned

s::Vector{Bool} Array of cells to be removed.

Returns FlowCytometryExperiment with the cells not prunned.

FlowCytometry.removeCells! — Function

function removeCells!(fcs::FlowCytometryExperiment, s::Vector{Bool})

Remove cells from the dataset.

Arguments:

fcs::FlowCytometryExperiment Dataset to be pruned

s::Vector{Bool} Array of cells to be removed.

Returns Nothing

FlowCytometry.removeChannels — Function

function removeChannels(fcs::FlowCytometryExperiment, s::Vector{Bool})

Remove channels from the dataset and return a copy.

Arguments:

fcs::FlowCytometryExperiment Dataset to be pruned

s::Vector{Bool} Array of channels to be removed.

Returns FlowCytometryExperiment with the channels not removed.

FlowCytometry.removeChannels! — Function

function removeChannels!(fcs::FlowCytometryExperiment, s::Vector{Bool})

Remove channels from the dataset.

Arguments:

fcs::FlowCytometryExperiment Dataset to be pruned

s::Vector{Bool} Array of channels to be removed.

Returns FlowCytometryExperiment with the channels not removed.

FlowCytometry.renameControl! — Function

function renameControl!(fcs::FlowCytometryControl,oldName::String,newName::AbstractString)

Change the name of the a channel in the of a FlowCytometryControl object.

Arguments

fcs::FlowCytometryControl FlowCytometryControl where to change the name.

oldName::String Old name of the channel.

newName::AbstractString New name of the channel.

Returns Nothing

function renameControl!(fcs::FlowCytometryControl,dic::Dict{<:AbstractString,<:AbstractString})

Change the name of the channels in the of a FlowCytometryControl object.

Arguments

fcs::FlowCytometryControl FlowCytometryControl where to change the name.

dic::Dict{<:AbstractString,<:AbstractString} Dictionary of old names as keys and new names as values.

Returns Nothing

FlowCytometry.checkControlNames — Function

function checkControlNames(fcs::FlowCytometryControl)

function that checks that the keys of the FlowCytometryControl object correspond to channels defined in .var.channels.

Arguemtns

fcs::FlowCytometryControl FlowCytometryControl object where to check if consistent.

Returns Gives an error if there is any inconsistency.

IO

FlowCytometry.loadFCExperiment — Function

function loadFCExperiment(file::String)

Function that loads a ftc experiment into a FlowCytometryExperiment.

Arguments:

file::String: Path to .ftc file.

Returns: FlowCytometryExperiment with the data from the uploaded file.

FlowCytometry.loadFCControls — Function

function loadFCControls(dic::Dict{String,String})

Function to upload all the files as given by a dictionary of control files and the corresponding channels.

Arguments

dic::Dict{String,String} Dictionary with name_of_file=>name_of_controlled_channel.

Returns FlowCytometryControl object with all the controls uploaded.

FlowCytometry.saveH5fcs — Function

function saveH5fcs(fcs::FlowCytometryExperiment,file::String)

Function to save in .h5fcs a FlowCytometryExperiment.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment to save.

file::String File name where to store the data.

Returns Nothing

function saveH5fcs(fcs::FlowCytometryControl,file::String)

Function to save in .h5fcs a FlowCytometryControl.

Arguments

fcs::FlowCytometryControl FlowCytometryControl to save.

file::String File name where to store the data.

Returns Nothing

FlowCytometry.loadH5fcs — Function

function loadH5fcs(file::String)

Function to load .h5fcs files into data.

Arguments

file::String File name of the datafile.

Returns FlowCytometryControl or FlowCytometryExperiment loaded from file.

Gating

FlowCytometry.isInsideGate — Function

function isInsideGate(p::Tuple{<:Real,<:Real}, gate::FlowCytometryGate)

Function that given a point and a gate, checks which entries are inside the gate polygon.

Arguments:

p::Tuple{<:Real,<:Real} Point to check if inside polygon.

gate::FlowCytometryGate Gate to check if the rows are inside the polygon of the gate.

Returns: Bool. True entries are the ones inside the gate polygon.

function isInsideGate(x::Matrix{<:Real}, gate::FlowCytometryGate)

Function that given the entries of two channels and a gate checks which entries are inside the gate polygon.

Arguments:

x::Matrix{<:Real} Matrix with N cells and 2 columns (channels).

gate::FlowCytometryGate Gate to check if the rows are inside the polygon of the gate.

Returns: Vector{Bool} with N entries. True entries are the ones inside the gate polygon.

FlowCytometry.Gating.filterByGate! — Function

function filterByGate!(fcs::FlowCytometryExperiment,gate::FlowCytometryGate,addKey::String)

Function that checks which cells are inside a gate.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment where to compute the gate

gate::FlowCytometryGate Gate to filter the cells.

addKey::String Kay to be added in .obs

Returns Nothing. A columns with addKey is added to .obs.

function filterByGate!(fcs::FlowCytometryExperiment,gates::Vector{FlowCytometryGate},addKey::String)

Function that checks which cells are inside a set of gates.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment where to compute the gate

gates::Vector{FlowCytometryGate} Set of gates to filter the cells.

addKey::String Kay to be added in .obs

Returns Nothing. A columns with addKey is added to .obs.

function filterByGate!(fcs::FlowCytometryExperiment,gateName::String,addKey=string(gateName,"_gate"))

Function that checks which cells are inside a gate.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment where to compute the gate

gateName::String Name of gate inside FlowCytometryExperiment.

addKey::String=string(gateName,"_gate") Key to be added in .obs

Returns Nothing. A columns with addKey is added to .obs.

function filterByGate!(fcs::FlowCytometryExperiment,gateNames::Vector{String},addKey)

Function that checks which cells are inside a set of gates.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment where to compute the gate

gateNames::Vector{String} Name of gates inside FlowCytometryExperiment.

addKey::String=string(gateName,"_gate") Key to be added in .obs

Returns Nothing. A columns with addKey is added to .obs.

FlowCytometry.Gating.manualGating! — Function

function manualGating!(fcs::FlowCytometryExperiment;debug=false)

Launch an interactive page accessible at localhost to define gates manually.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment object where to define the gates.

Returns Nothing. Adds or removes the gates to the object in place.

FlowCytometry.Gating.automaticQC! — Function

function automaticQC!(fcs::FlowCytometryExperiment;
    channel1="FSC-A",channel2="SSC-A",
    trim::Tuple{<:Real,<:Real}=(0.01,0.99),
    maxtrim::Real=.05,
    densityBandwidth::Tuple{<:Real,<:Real}=(.5,.3),
    finalBandwidth=.1,
    heightFromMax=.3,
    subsample=2000,
    keyAdded="automaticQC",
    seed = 0)

Function that automatically detects the main peak corresponding to viable cells in a FCS/SSC scatterplot and adds a gate to the object. The method follows the heuristics of Roca et al with some small adaptations.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment where to compute the gate

Keyword arguments

channel1="FSC-A" Channel name in channels corrresponding to the forwad scatter

channel2="SSC-A" Channel name in channels corrresponding to the forwad scatter

trim::Tuple{<:Real,<:Real}=(0.01,0.99) Percentile range in the channel ranges to use in the analysis before finding maximums.

maxtrim::Real=.05 Minimum in the range channels where to accept maximum points. This avoids the peak present in some datasets at lower expressions (debris).

densityBandwidth::Tuple{<:Real,<:Real}=(.5,.3) Relative bandwidth to smooth the density plots.

finalBandwidth=.1 Relative bandwidth to smooth the density plots in the final step.

heightFromMax=.3 In the last step, heigh at global maximum where to draw the contour of the QC region. Points with more probability that that will be used to make the QC gate.

subsample=2000 Subsample of points used for density steps of the system. Set to 'nothing' for using all the points (this can be extremely slow), with relative low numbers of points the gate output are already very good.

keyAdded="automaticQC" Key added to uns and gate name.

seed = 0 Seed to make subsampling reproducible.

Returns Nothing. A gate of automatic control is added to gates.

Compensation (Spillover/Unmixing)

FlowCytometry.Compensation.computeCompensationMatrix! — Function

function computeCompensationMatrix!(fcs::FlowCytometryControl; error::AbstractFloat = 10E-5, nMaxiterations::Int=10)

Compute the compensation matrix as proposed by Roca et al.

Arguments

fcs::FlowCytometryControl FlowCytometryControl where to compute the spillover matrix.

Keyword arguments

error::AbstractFloat = 10E-5 Maximum error to stop the refinement of the spillover matrix.

nMaxIterations::Int=10 Maximum number of iterations of the refinement matrix. Set to 0 to compute the classical spillover matrix.

Results Nothing. Adds the compensation matrix to the FlowCytometryControl object.

FlowCytometry.Compensation.compensate! — Function

function compensate!(fcs::FlowCytometryControl)

Compensate all the control experiments with the compensationMatrix of the object

Arguments

fcs::FlowCytometryControl FlowCytometryControl object where to compute the compensation.

Returns Nothing

function compensate!(fcs::FlowCytometryExperiment,compensation::CompensationMatrix)

Compensate FlowCytometryExperiment experiment with the compensationMatrix from an FlowCytometryControl object.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment object where to compute the compensation.

compensation::CompensationMatrix CompensationMatrix object

Returns Nothing, compensated the matrix X from fcs.

Dimensionaity reduction

FlowCytometry.DimensionalityReduction.pca! — Function

function pca!(fct::FlowCytometryExperiment;
    maxoutdim::Int = 0,
    method::Symbol = :auto,
    variance_ratio::Float64 = 0.99,
    mean = nothing,
    key_added::String = "pca",
    key_used_channels::Union{Nothing,String} = nothing
    )

Perform pca analysis over the channels.

Arguments:

fct::FlowCytometryExperiment FlowCytometryExperiment to perform the pca.

Keyword Arguments

maxoutdim = 0 Number of PCS to compute if 0, it will compute NChannels-1

method = :auto Method to compute pca. Choose between :svd, :cov and :auto.

variance_ratio = 0.99 Variance ratio up to which compute pca.

mean = nothing If to consider the normalization as performed. nothing will consider the mean not adjusted.

key_added::String = "pca" Key to be added to object.

keyusedchannels::Union{Nothing,String} = nothing Bool column from .var specifying which channels to use for clustering.

Returns Nothing. To the FlowCytometryExperiment object provided, PCs are added to the .obsm[keyadded] and information of the algorithm if included in .uns[keyadded].

FlowCytometry.DimensionalityReduction.umap! — Function

function umap!(fct::FlowCytometryExperiment;
    n_dims::Int = 2,
    key_added::String = "umap",
    key_obsm::Union{Nothing,String} = nothing,
    n_components::Union{Nothing,Int} = nothing,
    key_used_channels::Union{Nothing,String} = nothing,
    kwargs...
    )

Function that computes the UMAP plot of the neighbors.

Arguments

fct::FlowCytometryExperiment FlowCytometryExperiment to perform the UMAP.

Keyword Arguments

n_dims::Int = 2 Number of dimensions for the projection.

key_added::String = "KMeans" Key to be added to object.

key_obsm::Union{Nothing,String} = nothing Matrix in .obsm to use for the clustering. Only keyobsm or keyused_channels can be specified.

n_components::Union{Nothing,Int} = nothing Number of components to use from the .obsm.

keyusedchannels::Union{Nothing,String} = nothing Bool column from .var specifying which channels to use for clustering.

kwargs... Keyword argumetns for the UMAP_ function. Do ?FlowCytometry.UMAP_ for more details on the keyword arguments.

Clustering

<!– Clustering.kmeansTuning Clustering.agglomerative! –>

FlowCytometry.Clustering.kmeans! — Function

function kmeans!(fct::FlowCytometryExperiment;
    k::Int = 2, 
    metric = Distances.SqEuclidean(0.0),
    init::Symbol = :kmpp,
    key_added::String = "kmeans",
    key_obsm::Union{Nothing,String} = nothing,
    n_components::Union{Nothing,Int} = nothing,
    key_used_channels::Union{Nothing,String} = nothing)

Function that clusters the data using the KMeans algorithm. To help asses the number of clusters you can use KMeansTuning function.

Arguments

fct::FlowCytometryExperiment FlowCytometryExperiment to cluster the data.

Keyword Arguments

k::Int Number of clusters to partition the data.

metric = Distances.SqEuclidean(0.0) Metric to compute distances.

init::Symbol = :kmpp Initialisation algorithm of the KMeans. Choose between :kmpp, :rand or :kmenc. ?Clustering.kmeans_ for more information.

key_added::String = "KMeans" Key to be added to object.

key_obsm::Union{Nothing,String} = nothing Matrix in .obsm to use for the clustering. Only keyobsm or keyused_channels can be specified.

n_components::Union{Nothing,Int} = nothing Number of components to use from the .obsm.

key_used _channels::Union{Nothing,String} = nothing Bool column from .var specifying which channels to use for clustering.

Returns Nothing. To the FlowCytometryExperiment object provided, clusters identities are added to the .obs[keyadded] and information of the algorithm if included in .uns[keyadded].

Missing docstring.

Missing docstring for Clustering.gaussianMixture!. Check Documenter's build log for details.

Ploting

FlowCytometry.FCSPloting.plotQCSteps — Function

function plotQCSteps(fcs::FlowCytometryExperiment,key::String="automaticQC")

Function that returns the ploting steps from automaticQC.

Arguments

fcs::FlowCytometryExperiment FlowCytometryExperiment object where to define the gates.

key::String="automaticQC" Key used to save the results from the automatic control object.

Returns Figure with the steps ploted.

FlowCytometry.FCSPloting.plotControls — Function

function plotControls(fcsControl::FlowCytometryControl,channels::Vector{Tuple{String,String}})

Function that makes an scatterplot of the tuples of channels among controls and channels to see the spillover. If computeCompensationMatrix! has been computed, it also plots the correction after the control.

Arguments:

fcsControl::FlowCytometryControl FlowCytometryControl object where to check the spillover.

channels::Vector{Tuple{String,String}} Vetor of Tuples of control sample and spilled channel that want to be seen.

Returns List of figures with all the scatterplots between tuples. It can be ploted using plot(returned...,layout=(...))

FlowCytometry.FCSPloting.plotControlsCorrelations — Function

function plotControlsCorrelations(fcsControl::FlowCytometryControl)

Function that displays a heatmap with the correlation between controls and channels. It can work as a summary of spillover between data.

Arguments:

fcsControl::FlowCytometryControl FlowCytometryControl object to compute the correlations between channels.

Return Heatmap figure with the control channels in rows and channels in columns.